backend/requirements.txt
fastapi
uvicorn
google-generativeai
langchain
faiss-cpu
requests
python-dotenv
python-multipart
pypdf
python-docx
mysql-connector-python
sentence-transformers
beautifulsoup4
lxml
feedparser